backend/app/requirements.txt
Django>=3.0
psycopg2>=2.8
djangorestframework==3.14.0
pyjwt==2.8.0